change to follow naming convention on mixin classes
