+++ patchwork-enum-hacks/src/main/java/com/patchworkmc/mixin/enumhacks/MixinRarity.java
@@ -24,19 +24,19 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.Shadow;
 
-import net.minecraft.structure.pool.StructurePool;
+import net.minecraft.util.Rarity;
 
 import com.patchworkmc.impl.enumhacks.HackableEnum;
 
-@Mixin(StructurePool.Projection.class)
-public class StructurePoolProjectionMixin implements HackableEnum<StructurePool.Projection> {
+@Mixin(Rarity.class)
+public class MixinRarity implements HackableEnum<Rarity> {
 	@Shadow
 	@Final
 	@Mutable
-	private static StructurePool.Projection[] field_16683;
+	private static Rarity[] field_8905;
 
 	@Override
-	public void patchwork_setValues(StructurePool.Projection[] values) {
-		field_16683 = values;
+	public void patchwork_setValues(Rarity[] values) {
+		field_8905 = values;
 	}
 }
